move tests to disk graph
